refactor: change Vector2D from float to int and update serialization methods accordingly

+++ common/utils/vector_2d.h
@@ -6,16 +6,16 @@
 
 class Vector2D {
 private:
-    float x;
-    float y;
+    int x;
+    int y;
 
 public:
-    Vector2D(float x, float y): x(x), y(y) {}
+    Vector2D(int x, int y): x(x), y(y) {}
     Vector2D(const Vector2D& v1, const Vector2D& v2):
             x(v2.get_x() - v1.get_x()), y(v2.get_y() - v1.get_y()) {}
 
-    float get_x() const { return x; }
-    float get_y() const { return y; }
+    int get_x() const { return x; }
+    int get_y() const { return y; }
 
     Vector2D operator+(const Vector2D& other) const { return Vector2D(x + other.x, y + other.y); }
